Run thru initialization

+++ libs/spirit1/SPIRIT1_Library/Src/SPIRIT_Types.c
@@ -129,7 +129,7 @@ volatile SpiritStatus g_xStatus;
 void s_assert_failed(uint8_t* file, uint32_t line)
 {
   /* User can add his own implementation to report the file name and line number */
-  printf("Wrong parameters value: file %s on line %d\r\n", file, line);
+  printf("Wrong parameters value: file %s on line %d\r\n", file, (int)line);
 
   /* Infinite loop */
   while (1)
